pyproject.toml
[tool.poetry]
name = "memetext"
version = "0.1.1"
description = ""
authors = ["Ryan Townshend <citizen.townshend@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Pillow = "^8.2.0"
click = "^8.0.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
memetext = "memetext.memetext:main"